requirements.txt
attrs==24.2.0
black==24.8.0
blinker==1.8.2
Brotli==1.1.0
cachelib==0.13.0
certifi==2024.8.30
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.3.0
cssselect2==0.7.0
cycler==0.12.1
distlib==0.3.8
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Session==0.8.0
fonttools==4.53.1
gevent==24.2.1
geventhttpclient==2.3.1
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
identify==2.6.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
locust==2.31.6
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.9.2
msgpack==1.1.0
msgspec==0.18.6
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
networkx==3.3
networkz==1.0.6
nodeenv==1.9.1
numpy==2.1.1
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.3
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pre-commit==3.8.0
psutil==6.0.0
pycparser==2.22
pydyf==0.11.0
pyparsing==3.1.4
pyphen==0.16.0
PySocks==1.7.1
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
ruff==0.6.7
seaborn==0.13.2
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
tenacity==9.0.0
tinycss2==1.3.0
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.26.5
weasyprint==62.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.4
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
zopfli==0.2.3
